pyproject.toml
[tool.poetry]
name = "menu-app"
version = "0.1.0"
description = ""
authors = ["Allen8Kane <allen8kane@gmail.com>"]
readme = "README.md"
packages = [{include = "menu_app"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django = "^4.2.5"
markdown = "^3.4.4"
django-filter = "^23.2"
django-cors-headers = "^4.2.0"
pillow = "^10.0.0"
djoser = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
